DeleteElemeny

+++ Recursion/Source.cpp
@@ -13,7 +13,7 @@ void elevator(int floor)
 	cout << "Вы на " << floor << " этаже" << endl;
 }
 
-int factorial(int n)
+long long int factorial(int n)
 {
 	
 	if (n == 0 || n==1) return 1;
